# services/trading-service/docs/swagger.yaml
definitions:
  dto.AssetType:
    enum:
    - STOCK
    - FUTURES
    - OPTION
    - FOREX
    type: string
    x-enum-varnames:
    - AssetTypeStock
    - AssetTypeFutures
    - AssetTypeOption
    - AssetTypeForex
...
  dto.ExerciseOptionRequest:
    properties:
      account_number:
        type: string
    required:
    - account_number
    type: object
  dto.ExerciseOptionResponse:
    properties:
      destination_amount:
        type: number
      destination_currency_code:
        type: string
      exercised_contracts:
        type: integer
      option_asset_id:
        type: integer
      purchased_shares:
        type: number
      remaining_contracts:
        type: integer
      remaining_option_shares:
        type: number
      source_amount:
        type: number
      source_currency_code:
        type: string
      stock_asset_id:
        type: integer
      strike_price:
        type: number
      total_cost:
        type: number
    type: object
...
  dto.PortfolioAssetResponse:
    properties:
      amount:
        type: number
      asset_id:
        type: integer
      avg_buy_price_rsd:
        type: number
      last_modified:
        type: string
      price_per_unit_rsd:
        type: number
      profit:
        type: number
      public_amount:
        type: number
      ticker:
        type: string
      type:
        $ref: '#/definitions/dto.AssetType'
    type: object
  dto.PortfolioProfitResponse:
    properties:
      total_profit_rsd:
        type: number
    type: object
  dto.PublishAssetRequest:
    properties:
      amount:
        type: number
    required:
    - amount
    type: object
...
  errors.AppError:
    properties:
      message:
        type: string
      status:
        type: string
      timestamp:
        type: string
    type: object
...
info:
  contact: {}
  description: API for managing portfolios, executing trades, and handling financial
    market operations.
  title: Trading Service API
  version: "1.0"
paths:
...
  /api/actuary/{actId}/assets:
    get:
      description: Returns all currently held asset positions for an actuary (employee
        agent/supervisor), aggregated from all orders. Only approved orders with fills
        are counted. Includes stocks, futures, options, and forex pairs.
      parameters:
      - description: Actuary ID
        in: path
        name: actId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PortfolioAssetResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Get portfolio for an actuary/agent
      tags:
      - portfolio
  /api/actuary/{actId}/assets/{ownershipId}/publish:
    patch:
      consumes:
      - application/json
      description: Appends the number of assets the caller makes publicly visible
        on the OTC portal.
      parameters:
      - description: Asset ownership ID
        in: path
        name: ownershipId
        required: true
        type: integer
      - description: Amount to make public
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PublishAssetRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Publish assets for OTC trading
      tags:
      - otc
  /api/actuary/{actId}/assets/profit:
    get:
      description: Returns the total accumulated profit across all currently held
        asset positions for an actuary.
      parameters:
      - description: Actuary ID
        in: path
        name: actId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PortfolioProfitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Get total profit for an actuary portfolio
      tags:
      - portfolio
  /api/actuary/{actId}/options/{assetId}/exercise:
    post:
      consumes:
      - application/json
      description: Exercises one contract of an actuary-owned in-the-money call option
        and buys the underlying stock at the strike price.
      parameters:
      - description: Actuary ID
        in: path
        name: actId
        required: true
        type: integer
      - description: Option asset ID
        in: path
        name: assetId
        required: true
        type: integer
      - description: Exercise request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ExerciseOptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ExerciseOptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Exercise an owned option
      tags:
      - portfolio
...
  /api/client/{clientId}/assets:
    get:
      description: Returns all currently held asset positions for a client, aggregated
        from all orders. Only approved orders with fills are counted. Includes stocks,
        futures, options, and forex pairs.
      parameters:
      - description: Client ID
        in: path
        name: clientId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PortfolioAssetResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Get portfolio for a client
      tags:
      - portfolio
  /api/client/{clientId}/assets/{ownershipId}/publish:
    patch:
      consumes:
      - application/json
      description: Appends the number of assets the caller makes publicly visible
        on the OTC portal.
      parameters:
      - description: Asset ownership ID
        in: path
        name: ownershipId
        required: true
        type: integer
      - description: Amount to make public
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PublishAssetRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Publish assets for OTC trading
      tags:
      - otc
  /api/client/{clientId}/assets/profit:
    get:
      description: Returns the total accumulated profit across all currently held
        asset positions for a client.
      parameters:
      - description: Client ID
        in: path
        name: clientId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PortfolioProfitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Get total profit for a client portfolio
      tags:
      - portfolio
...
  /api/otc/public:
    get:
      description: Returns a paginated list of assets that have been marked public
        by their owners on the OTC portal.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 10)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: List all publicly available OTC assets
      tags:
      - otc
...
securityDefinitions:
  BearerAuth:
    description: JWT Authorization header using the Bearer scheme.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"